move logo to cli
