+++ cli/logo.cjs
